refactor: move ListingTable over to the catalog-listing-table crate (#18080)
## Which issue does this PR close?

- This addresses part of
https://github.com/apache/datafusion/issues/17713
- Closes https://github.com/apache/datafusion/issues/14462


## Rationale for this change

In order to remove the `datafusion` core crate from `proto` as a
dependency, we need to access `ListingTable` but it is within the `core`
crate. There already exists a `datafusion-catalog-listing` which is bare
and appears to be the place this should exist.

## What changes are included in this PR?

Move `ListingTable` and some of its dependent structs over to the
`datafusion-catalog-listing` crate.

There is one dependency I wasn't able to remove from the `core` crate,
which is inferring the listing table configuration options. That is
because within this method it downcasts `Session` to `SessionState`. If
a downstream user ever attempts to implement `Session` themselves, these
methods also would not work. Because it would cause a circular
dependency, we cannot also lift the method we need out of `SessionState`
to `Session`. Instead I took the approach of splitting off the two
methods that require `SessionState` as an extension trait for the
listing table config.

From the git diff this appears to be a large change (+1637/-1519)
however the *vast* majority of that is copying the code from one file
into another. I have added a comment on the significant change.

## Are these changes tested?

Existing unit tests show no regression. This is just a code refactor.

## Are there any user-facing changes?

Users may need to update their use paths.
