+++ datafusion/core/src/datasource/dynamic_file.rs
@@ -20,6 +20,7 @@
 
 use std::sync::Arc;
 
+use crate::datasource::listing::ListingTableConfigExt;
 use crate::datasource::listing::{ListingTable, ListingTableConfig, ListingTableUrl};
 use crate::datasource::TableProvider;
 use crate::error::Result;
