+++ datafusion/core/src/datasource/listing/mod.rs
@@ -20,7 +20,8 @@
 
 mod table;
 pub use datafusion_catalog_listing::helpers;
+pub use datafusion_catalog_listing::{ListingOptions, ListingTable, ListingTableConfig};
 pub use datafusion_datasource::{
     FileRange, ListingTableUrl, PartitionedFile, PartitionedFileStream,
 };
-pub use table::{ListingOptions, ListingTable, ListingTableConfig};
+pub use table::ListingTableConfigExt;
